requirements.txt
streamlit>=1.45.1
pandas>=2.2.3
numpy>=2.2.6
matplotlib>=3.10.3
seaborn>=0.13.2
gspread>=5.12.4
gspread-pandas>=3.3.0
google-auth>=2.40.2